noria-client: Implement prepared DELETEs to noria
Allow preparing and executing DELETE statements directly to Noria when
no fallback is configured in an adapter. This is mostly just a copy of
the same code for UPDATE, with all the stuff around parameters for
columns removed.

This isn't something a *customer* will likely ever interact with, but is
useful when testing things without fallback (in my case I'm using it via
the vertical test suite).

Fixes: ENG-169
Change-Id: I06d2813634a2df160d7ce21a67193b5530dba783
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/734
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>