vertical-tests: Add test for a simple inner join
Add a second query to the vertical-test suite for a simple two-table
inner join with a partial key on the join key. This required some
alterations to the vertical test harness to get working:

- despite what I was led to believe about macro hygiene, we do in fact
  need to generate a separate name for the `generate_ops` function we
  make as part of the macro, as separate invocations will hit a
  duplicate function error - this uses the `paste` crate to concat
  together identifiers for the name of the function.
- When collecting keys from previously generated rows, it might be the
  case that the tables the keys are from have not had any rows generated
  yet - previously we'd unwrap and panic here, now we just filter those
  out.

One thing that is *not* happening in this commit but should happen soon
is making it more likely that we're going to generate actually matching
foreign keys for these joins, as currently it's diminishingly unlikely
that the values on either side will ever actually match up. That's
coming next.

Refs: ENG-674
Change-Id: I485bb1bb923179ebc6e2b1feb1d6e3ffe407d1d0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/739
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>