[Live QCA] Implement QueryStatusCache to track allow/deny lists.
This CL adds the `QueryStatusCache` that will store the status
of each query we have seen in our system. The adapter will use the
cache to determine the set of queries we don't support and execute
them directly against fallback.

This cache will be used by the adapter, but managed by a separate
async query reconciler, which will issue queries to upstream and Noria
to determine if we support the query based on, successful execution and
column header matching.

The `QueryStatusCache` and the remaining live query coverage work will
be placed behind the `live_qca` compile flag.

Co-Authored-By: Justin Miron<justin@readyset.io>
Change-Id: I026cd89b43d03b650e428e6d3380698a8a43660b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/762
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>