Support new per-query histograms in the adapter.
This adds logging histogram metrics per-query in the adapter,
QUERY_LOG_EXECUTION_TIME. This will give us histograms per-query
in the adapter. This would be extremely fine-grained for
ad-hoc metrics, as such, the query text is not logged. Once
we support anonymizing and removing literals from ad-hoc queries
this can be changed.

To keep this new logging off the critical path, logging
is done with a separate task and is passed the QueryExecutionEvent
at the end of query execution.

Change-Id: If3cc792a40f3f1a641dc4fba24ec7d2c251362b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/743
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>