Refactor adapter metrics path to use `QueryExecutionEvent`.
The query-coverage-analysis work laid the groundwork for updating
an event during the execution of a query (while it is enabled).
This can be useful for recording per-query metrics and understanding
what errors were thrown by what queries - right now logs only
indicate that noria threw an error for *some* query.

This is initially disabled when --mirror-reads is not passed to limit
impact on performance. However, the performance of updating a single
struct during the lifecycle of a query is incredibly cheap compared
to the myraid of things we do during query execution: fetching
views, query parsing, serialization, deserialization. The benefits:
being able to log better per-query metrics/errors, and makes the
code path much simpler.

This refactors some of the QCA metrics path into the main code path
as it will be used for both QCA and improve performance metrics for
the live QCA work.

Change-Id: I0c20ecc0583ff9a280f4c95598850b4574c4a344
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/742
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>