Prepare statements to both noria and upstream.
We currently only prepare to one of noria and upstream
when we receive a prepare statement. When this is
noria we attempt to on-the-fly prepare the statement
if noria requires fallback for some reason.

It makes more sense to issue prepare statements to both
noria and upstream. This is captured in the implementation
by making each prepared statement be associated with
an id in both noria and upstream.

This functionality also lets us fallback easier and
issue the same query against both Noria and Mysql
in the future, which is something the query reconciler
requires.

Change-Id: I9bd83dc8c01fa53d458fec839e94e57c91b883ce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/744
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>