vertical-tests: Generate values for foreign keys
For joins, we'd like to make it reasonably likely that we're actually
going to match on the join key, as otherwise we're basically never going
to return results. This accomplishes that by swapping out the general
"give me a random value of this type" strategy for foreign key columns
with a strategy that waits to generate rows for tables with foreign keys
until after we've generated tables that they have foreign keys to, and
then picks a random value from the column in the target table to
generate the value. In the `vertical_tests!` macro, this is represented
by replacing the type of the column with something that looks like
`foreign_key("users", 0)`.

Change-Id: Ib738554fbbd53f4a3fa6ecc54d3d3ed86ec0621e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/747
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>