treewide: Convert eprintlns to proper log messages
Back when we were using slog, in order to do proper logging we had to
pass around an explicit logger object - this was enough of a pain (or
even impossible in the case of traits that didn't expect loggers as
arguments) that we regularly just used bare `eprintln` calls instead of
proper log messages. Now we're using tracing, which allows us to log
anywhere we like, so we should replace these eprintlns with actual
proper log messages.

Change-Id: I89f62d910c7d1a34c70c392e83d9860f0d5e8056
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/746
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>