nom-sql: Expand Visitor to cover the entire AST
Following the already-established pattern for SelectStatement, expand
`nom_sql::analysis::visit::Visitor` to cover the *entire* SQL AST,
starting at `SqlQuery` and descending. This is pretty huge, but
basically entirely mechanical.

I'm doing this right now because I'll want to use this to schema-qualify
tables, but this is also something we've been wanting for a while to
allow us to anonymize literals in queries other than `SelectStatement`s.

Change-Id: I62dc4d9d4e241f9f35839a0808a3799c5d7a4e40
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/754
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>