nom-sql: Define the start of an AST visitor
Add the beginnings of an implementation of an AST visitor pattern for
nom-sql's AST, currently only with implementation for all of the AST
rooted at SelectStatement. The specific approach here is directly
inspired by the AST visitor that rustc itself uses, which is linked in a
doc-comment in the code.

We have a *lot* of code that recursively traverses ASTs, and every time
we do that we have to re-implement the actual traversal from scratch,
which is both bug-prone (we might forget - and frequently have forgotten
- to traverse down some attribute of some AST node) and verbose. The
visitor pattern allows us to define the AST *traversal* once, decoupling
it from the actual logic of the algorithm that needs to traverse the
AST. This is demonstrated in tests via a doctest that counts the
occurrences of literals, and a unit test that counts AST nodes
overall (to give us a sanity-check that the visitor itself isn't
forgetting to traverse anything).

Change-Id: I7d88479af87336a86eda703a64088bc47cd4160a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/755
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>