substrate: Fix configuration of Packer IAM role
Turns out that if you do not include a name for the inline policy for
an IAM role, Terraform will claim to apply it and do nothing. This is
reported as a bug.

https://github.com/hashicorp/terraform-provider-aws/issues/19145

This has already been applied so the plan will be nothing.

Change-Id: Ic15b42fa9594a3c33316d95e85b9d1b9e30f692c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/758
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>