noria-client: Add function to autoparametrize query
Add a function to noria-client, implemented via a Visitor, that
replaces all literals in positions we support parameters (eg only
directly compared to columns on either side of an AND operator) with
parameters, and returns the set of replaced literals and their position
in the parameter list.

The plan here is to use this for all queries we receive, especially
ad-hoc unprepared queries, so that users who are using client libraries
that string-interpolate parameters rather than providing them as
parameters to prepared statements don't accidentally create huge fully
materialized queries that explode their ReadySet memory usage.

Refs: ENG-683
Change-Id: Idde3e3bbe0f4ec959a88a0fbe689213437d0253c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/765
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI