[Live QCA] Integration of QueryStatusCache with adapter serving path.
When `--live-qca` is passed to the adapter, a new serving path is
enabled for SelectStatement's. This will check the query status
in the QueryStatusCache. If the query is allowed, these queries
will then be executed on noria, falling back to upstream when
needed. If the query is still processing, NeedsProcessing, then
the query will be executed on fallback.

Since there is no mechanism for moving queries from NeedsProcessing
to Allowed, this currently will execute every query on fallback
when --live-qca is enabled.

Change-Id: I6a68ec2e1045723e4c637e76b37c64790ea32ab8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/768
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>