Propagate create options
Currently we drop the table create options, however I
need at least the auto-increment to deduce the number of
rows in a table for the data generator, and we will definitely
want collate and charset to properly support collation in the
future.

Not all the options are fully propagated, some are replaced by
a generic placeholder.

Change-Id: Ib4369e39e5b1f0c1bf0d65acd4689943404ef9e7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/770
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>