[Live QCA] Implement query reconciler thread in the adapter.
This introduces the query reconciler, a seperate thread
responsible for performing migrations and confirming that they
complete successfully. It manages the adapter's allow/deny list
stored in the `QueryStatusCache`. When queries are successfully
run on both noria and the upstream db they are placed in the
allow list, otherwise they are in the deny list until they are
able to be successfully run.

This first implementation treats correctness as: did the two
prepares succeed without an error in both. Future changes will
incorporate column header verification.

Change-Id: Id62579a16f4071b76bb42b5d93484c174b1d8391
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/772
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>