Vector setup script should fail if there is no consul aggregator yet.
Returning a status code of 1 from the vector agent's
setup-vector service script will prevent it from starting if there
was no registered consul aggregator service. Instead, it will loop
restarting the service and checking if an aggregator ha started.

This will ensure that whenever the vector agent service starts,
it will be attempting to connect to a valid aggregator.

Change-Id: I8f2f95747478cce084f83ef73843fef355edd49e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/773
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>