noria-client: Wrap up where-in and autoparametrization
Wrap up both the where-in collapse rewrite and auto-parametrization into
a single `process_query` function, which returns a new
ProcessedQueryParams struct which has a single method to build a list of
lookup keys from user-supplied params given both where-in collapsing and
auto-parameters.

Since auto-parameters are fixed down to only be Literals, this prevents
us from doing the thing we were doing before to differentiate between
column indices that we expose to the client as part of the prepared
statement API and the actual indices of the key columns in the Noria
View, which was to calculate the column indices pre-rewrite and then
apply explode_params to the vector of column indices. Instead, we now
calculate column indices both before and after rewrite and use the
before-rewrite version to get the param columns to send to the client
and the after-rewrite version to get the key column indices for noria.
This gives us an opportunity to wrap up the columns -> indices ->
ColumnSchema conversion into a single method on noria::View with much
better asymptotics than the previous version.

As of this commit, auto-parametrization is working and tested on
prepared statements sent through the adapter, but not on ad-hoc queries
yet - that's coming next.

Fixes: ENG-683
Change-Id: Ie53b6e1f906062152339ca65027964d4c3009d28
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/779
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>