Improve error message for misconfigured binlog
If binlog replication is misconfigured in the upstream MySQL database
the error message currently just says "Empty response for SHOW MASTER
STATUS"; recently, when debugging this, we learned that this can happen
if either the binlog_format is set to something other than ROW or if in
RDS the backup retention period is set to 0 days. This updates the error
message for this to suggest fixing those two things, to ease customer
debugging when this happens.

Change-Id: If4510cc61d64e3b4cc03eff5ff679bfa1bf36e7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/783
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI