logictests/requires-fallback/framework/php/laravel:  Additional cases
This adds without-migrations.test and schema-and-queries-only.test;
these are converted from a querylog of the Laravel framework test with
the installation and migration sections replaced with a mysqldump of a
bare freshly-installed Flarum.  schema-and-queries-only.test also
removes USE and SET NAMES statements.

without-migrations.test fails on USE statements; removing those reveals
failures on SET NAMES.

schema-and-queries-only.test fails on UNION SELECTs (lines 485 and 565).
Removing the UNION queries reveals a failure at the INSERT on line 683,
with the error message "Statement takes 1 parameters but 5 was
supplied."

Change-Id: Ib7957e80bbe09bf866267a3704b53d83ec25e69c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/817
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>