ops/cfn-deploy: initial work for cfn deployment
Inspired by the Makefile that Buildkite has for their stack, create a
program that will modify the CloudFormation template using information
from a Packer manifest and otherwise update other defaults as needed.

As a first step, parse a Packer manifest and create YAML which can be
copied and pasted into the right location.

Change-Id: I45ff3d661ea260a9c961866ecb7e33ce5e00488d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/788
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>