Check column header count and types in reconciliation.
This adds the column header check for correctness in the reconciler
loop. On every new query (and failed queries after a delay) the
reconciler will run the prepare statements against upstream and
noria and will check that the column headers returned match. If
they do not, then noria does not handle the query correctly and
we will add the query to the deny list after several failures.

This is implemented by adding a new trait `NoriaCompare`, that
each upstream database's StatementMeta must implement. It
requires each upstream to implement checking their prepare
statement's metadata against noria's passed as nom-sql.

Change-Id: I8e9fbaf93e4d44fcf70226a068d8a02337693d1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/798
Reviewed-by: Peter Farr <peter@readyset.io>
Tested-by: Buildkite CI