noria_client: Disallow DISTINCT with parametrized IN
For the same reason we don't currently support aggregates with
parametrized IN, we also can't currently support DISTINCT with
parametrized IN, because we don't re-aggregate when doing multi-key
lookups. This is tracked at ENG-629.

This commit properly classifies queries with DISTINCT and parametrized
IN as unsupported, and moves the 54 tests that do those operations to
out-of-scope.

Change-Id: Ia1b5f9ed9eac9cc6491a2eac5db6397653f32e5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/806
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>