migration: Split materialization into 3 phases
Split the materialization phase of migration into 3 phases: extend,
validate, and commit. The first phase, which already existed as a
private function called by commit but is now exposed and called directly
by the outer migration function, extends the materializations with a set
of new nodes and sets up indexing obligations for everything. The second
validate phase has been broken out of commit and is responsible for
validating all the various graph invariants (most notably in this
particular case the rule forbidding partial nodes above full nodes). The
third phase, commit, is where the mutation of the DomainMigrationPlan
happens to actually inform domains about new nodes etc.

The idea here is that as part of the work to fix the partial-above-full
case, we're going to want to run `extend` and `validate` multiple times,
but only `commit` once (since we only want the final materializations
ending up in the DomainMigrationPlan)

Refs: ENG-472
Change-Id: If1e9ab682872ddf5908020dd11ea1fcc8e24138b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/807
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>