Add http router for adapter to expose metrics
This commit adds an http router for the adapter that can be used to
retrieve the current allow and deny lists, as computed when live qca is
enabled.

This builds the adapter router in the same way that the
NoriaServerHttpRouter is built out.

This also adds a new wrapper type around Vec<SelectStatement> for the
returned allow and deny lists. This was so we can have a custom
Serialize implementation for the wrapper type QueryList, which enforces
anonymization during serialization, as well as calling query.to_string()
for each given SelectStatement so we get human readable SQL strings to
present in our respective allow and deny lists.

Change-Id: Ia433a9dc6844f88387723092933ea4ddd01919b2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/811
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>