Error handling for noria-mir.
This commits introduces error handling for the `noria-mir` module.
The error handling changes were:
1. Change the signature of the `insert_column` method in
`noria/server/mir/src/node/node_inner.rs` to return a `ReadySetResult<()>`.
The whole chain of methods calling this one have been changed in the same way as
well.
2. Modified the code in `noria/server/mir/src/reuse.rs` to be more readable and
documented (and allowed) the use of an `unwrap`.
3. Modified the `column_specifications` method in `MirNode` to return a
`ReadySetResult`, and documented the behaviour.
4. Added a `#![warn(clippy::panic)]` to `mir/lib.rs`, to avoid reintroducing
panics in the future.

Refs: ENG-298
Change-Id: I4bcb05c84afac976a90d9299c12db22eb82681d6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/822
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>