unbounded-interval-tree: Balance on deletion
AVL-rebalance the unbounded interval tree after deleting nodes.

This was *devilishly* difficult to get right, but ended up being
astoundingly little code - essentially, after we finish walking a node
for deletion, fully rebalance it by first doing one of the four AVL
rotation cases (left-left, left-right, right-right, or right-left) then
recursively rebalancing the child in the direction that we just rotated
the top node, in a loop until the node becomes balanced.

I *think* this isn't as asymptotically terrible as it sounds, as since
in the worst case where we've just removed one of the children of the
root node entirely we would be imbalanced by a factor of the height of
the tree, which is log_2 of the number of intervals, so that should
ideally bound the number of rotations we have to do in that loop. We'll
have to benchmark this, though my instinct tells me that for our use
case this is going to now be dominated by the constant factor rather
than the asymptotic performance.

Regardless, this all comes with a fairly rigorous empirical
demonstration that the tree is always balanced after a deletion - I've
updated the `check_invariants` function to remove the parameter to check
balance, instead *always* checking that all nodes in the tree are
balanced, and run the `remove::preserves_invariants` property test
locally at 100k iterations without it failing. Also, the bit where left
and right rotation ended up not actually being the inverse of each other
has been removed, since per the old comment in the code that was only
the case because we needed to be able to insert into an unbalanced tree
- similarly, the *test case* to insert into an unbalanced tree is gone
now

Co-Authored-By: Dan Wilbanks <dan@readyset.io>
Change-Id: Idd96017e574c74497eca604dae3322f49d5d5f3d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/823
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>