Revert "Bump chrono to fork with time 0.3 as a dep" and ignore `chrono` issue.
This reverts commit d30386deb2dd83809b1996aab555c5b4cbacee8a.

Reason for revert: CI is failing because of this, plus this doesn't really fix
the vulnerability issue, since the issue is rooted in `chrono`, not in
`time`[0].
This is a temporary fix to unblock CI.

This commit is a revert + a change because we are at a deadlock here, where we
can't build the CL to ignore the `chrono` vulnerability because of the `chrono`
version bump failing; and we can't revert the `chrono` version bump (in a single
atomic commit that passes CI at least) without cargo audit failing because of
the vulnerability.

We also revert the changes to our fork of rust-postgres[1]

[0]: See https://github.com/chronotope/chrono/issues/602#issuecomment-940445390
for more information.
[1]: https://github.com/readysettech/rust-postgres/commit/278f24

Change-Id: I4424d929adb96138bbe0c9021ce12a017b2040e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/841
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>