Update health check to request from adapter's http endpoints.
The current health check creates and breaks a tcp connection
on the adapter's sql port. Since this doesn't perform the
correct handshaking, the adapter determines this as a failed
sql client and logs it as such. To remove this, we create a
new port http endpoint /health that the healthchecks will
use to verify liveness.

Change-Id: Ic229872f5a0d8f84d122b37774d1894eb3a89d2e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/828
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>