[CVE-2020-26235] Remove the usage of `chrono` unsafe methods.
We are removing the usages of `chrono::Local::now()` and
`chrono::Local::today()` methods, which call `localtime_r` under the hood, which
is unsafe[0].
For the `Literal` to `DataType` transformations, returning an error instead was
more difficult than expected, because it could be the case that a table has a
`CURRENT_TIMESTAMP()` as default value, in which case we transform the `Literal`
into `DataType`s at the moment of creation of the base tables; which is wrong by
design.
So, instead of returning an error and in order to preserve the logic and
minimize the number of changes to be made to become compliant, we are creating
the `NaiveDateTime` (needed for the `DataType::Timestamp`) using the `time`
library (version `0.3`, which is compliant); but using UTC instead of local
time. While it's true that MySQL uses local time, as previously mentioned, the
behaviour of these literals is already wrong, so using UTC for now is a
good-enough fix to get rid of the troublesome `chrono` methods.

We also want to keep the list of troublesome methods as a lint in clippy, so the
builds can fail if we ever introduce them again[1][2].
The list of methods include the `chrono` methods as well as the `time` methods.
The latter is to prevent engineers from accidentally using those methods with a
library version prior to 0.2.23.
If `time` methods are going to be used, then an explicit
`#[allow(clippy::disallowed_method)]` should be used and the usage must be
documented.

[0]: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-26235
[1]: https://rustsec.org/advisories/RUSTSEC-2020-0071
[2]: https://rustsec.org/advisories/RUSTSEC-2020-0159

Refs: ENG-702
Change-Id: Ic2df3a19d20120a8a69023f1d2c6a7ed6f9be274
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/835
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>