Remove obsolete files from previously open-source repos
A bunch of files were left over from the previous open-source libraries
that we forked and subtree'd in to this repo, including obsolete CI
config files in several places and licenses (which are no longer
accurate as our version of all of this code is proprietary). This
removes all of them.

Change-Id: I2973e4332deed56b0df8d3e049cb9de3875286ce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/874
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>