server: Don't enable sharding with shards set to 1
Sharding a graph with a shard factor of 1 makes very little sense - so
little sense, even, that I'd argue the *expected* user behavior is that
setting shards to 1 would actually disable sharding. This implements
that, in the same way that setting shards to 0 also disables sharding,
both in the Builder and in the interpretation to the command-line args.

Refs: ENG-703
Change-Id: I0b800aece3b53668ce3bf614e830c3810eb6ea9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/880
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>