Make replay paths IndexType-aware
Update all data structures and logic surrounding replay paths to use a
full `Index`, rather than just a set of columns, both in the
materialization planner and in the domain as part of execution of
dataflow. When building replay paths, the index type is threaded down
from where it's now being passed in by the reader, and then everywhere
we were previously using a set of columns to look up a replay path we
now use both the columns and the index type. In the latter case,
especially for internal upqueries (the upqueries we make to fill
internal state as part of executing another replay, eg for a join) the
index type in question *might* come from a KeyComparison, rather than an
Index directly. This leads to one potentially surprising part of this
change, which is that some things have *not* changed to use full indexes
instead of columns - notably the `ColumnMiss` and `ColumnRef` structures
still just store a set of keys and the columns those keys are on. This
is because the actual *selection* of which index needs to be used needs
to actually look at the replay paths available, and then use the keys to
make that selection - but where ColumnMiss and ColumnRef are used, which
is to remap upqueries for eg straddled joins, doesn't know about the
indices that are available yet. The data structure previously known as
OptColumnRef, on the other hand, is used to actually build
materialization plans for nodes, so *does* need to store a full Index -
so to clarify the difference between those two data structures I've
renamed OptColumnRef to IndexRef.

Also, since due to the way we're picking index types it is now *not*
allowed to make a point query into a btree-map-backed reader, this
commit also updates the `range_upquery_after_point_queries` integration
test to make both a btree and a hashmap reader and do the appropriate
queries to each. In the case of actual SQL this isn't an issue, since
every reader will only ever serve one query.

Change-Id: I3f3117a8fa86bad91c63adcea25040104d2304cb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/884
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI