topk: Bring back sorting by group
This partially reverts commit 2cf40b3eadd801ff46f70431754ccd314092af45,
which removed the behavior in both topk and join where we sort by the
key before processing records. In join this is just an optimization to
deal with negatives and positives for the same record, but in topk this
is actually essential to the operation of the node - this comes with a
unit test for the operator which was failing before this change.

Change-Id: I9a9d8cb90e5075fa88e429a614b0cc2e05a9a6e4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/886
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI