Support using fallback when noria backend cannot be created.
In the noria-adapter when a client connects, a backend is created.
If backend creation fails due to failure to create a noria backend,
we return an error to the client and the client is unable to issue
any queries.

This can happen if:
  * Connecting to noria takes longer than the tcp timeout set,
    though this is unlikely.
  * There is no live noria-server.
  * There is currently no controller while the system recovers
    from a failure.

In any of these cases, we can and should use fallback to handle
queries. This CL returns an error during backend creation instead
of panic'ing and failing the connection to the client. Subsequent
calls to the noria connector requiring the backend, will return
an error, allowing queries to go to fallback.

Change-Id: Id00c021436911f58b004ae9cfd4653f779e0135d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/895
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>