Add metadata to node metrics exported via node-exporter.
This adds per-node metadata to each metric exported via the
node-exporter through vector's remapping capabilities. This
allows the metrics to include information passed via
command-line arguments.

Each metric is typically logged into prometheus with the
source address and app, but since we use a metrics aggregator
this information returns the vector aggregator's address,
and not the server/app the metrics came from.

Change-Id: I139d301451b9736630cd0958f605ee86fbbbf019
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/898
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>