dataflow: Make ingredient::lookup return an enum
This horrible nested `Option<Option<Box<dyn Iterator<...` mess that was
the return type of Ingredient::lookup was clearly a bad idea, as we both
had to explicitly document it, and clippy was complaining about it via
the type_complexity lint we had to explicitly allow.

That mess, to restate the previous docs, was:

- None, if there was no materialization
- Some(None), if there was a miss
- Some(Some(records)), if there were records

Of those, the first is *never* a case that we need to do anything
special to handle, so we can just make the whole method return a
ReadySetResult of either records or miss, and have the "you did a lookup
into an index that doesn't exist" case be a new ReadySetError variant.
For the remaining two, this defines a new IngredientLookupResult enum,
which looks a lot like the underlying LookupResult enum which State
uses, except the iterated-over values are themselves ReadySetResults,
which allows for failure in query_through in cases like expression
evaluation in both filter and project.

This is a refactor I've been wanting to do for a long time (again, the
type was really ugly) but is happening now as I work on index selection
as this will make it a lot simpler to thread through type information
that's necessary to deal with lookups into different indexes.

Change-Id: Idda7e2581d48dd489fbd93999a30c827e8a6bfc2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/901
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>