topk: Don't consider positives when we're already at k
If a topk node which has k records receives a positive delta for a group
which is below the minimum element within the k records, it's possible
that there are *other* records in the parent between the minimum element
and that positive record. If we add that record to the group at the same
time as we delete some other record, we would then return incorrect
results due to not considering those rows in the base table between our
minimum element and the newly received records. This can happen when eg
updating a record within k such that it moves out of k when some other
record should be inserted into k.

Instead, this change just discards all positives received by a topk
below its minimum record within that group, if the group is already
at (or above, for good measure, though this should never happen) k
records. If we also receive deletes in the batch that drop us below k
we'll just query our parent to backfill during post_group.

To make the asymptotics here nicer, this changes `current` (internal to
Topk::on_input) to use a BinaryHeap, which stays sorted on insertion and
provides cheap (amortized O(1)) methods to retrieve the current minimal
element. To make that work we have to define a temporary CurrentRecord
struct with a reference to the `order` of the topk as a field, since
annoyingly none of the rust stdlib data structures support external
comparators.

Fixes: ENG-709
Change-Id: I92f5eb131c7c50b2a2f1382beeb9b7142da73e97
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/905
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>