Add metrics aggregator Dockerfile
This commit adds a Dockerfile for building a container of the
metrics-aggregator binary. I have also expanded the Makefile to include
building this docker container, and pushing it to our registry.

This has been tested locally and does successfully create a
metrics-aggregator container as intended.

This also adds a deploy folder into the metrics aggregator crate for
local testing. This deploys noria-server, noria-mysql-adapter, consul,
prometheus server (set to scrape the adapter), and the metrics
aggregator.

Change-Id: Ic7fb72081d8fe90b75858ae54a3c20e31a8eba64
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/908
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI