Add `query-validate` feature flag to adapter.
When `query-validate` is enabled, the adapter will verify that
the query executes in both noria and upstream. Ad-hoc queries
are explicitely prepared before execution to verify that their
schemas match.

This can uses the same code path as the reconciler to verify
that queries prepare in both upstream and noria. This can be
extended to use schema comparison in the future.

Change-Id: I80ee0fe1c3a4166fb0ba38598fe7c3cdfac01984
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/912
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>