topk: Remove record cache
Now that topk can query its parent to backfill up to k records when
deletes bring it below k, there's no need to maintain a separate cache
of records internal to the topk in the case of full materialization - we
can just indiscriminately use the same query-the-parent mechanism for
full materialization. This removes the extra_records cache from topk and
everything that was around to support it, updates the unit tests to
properly materialize into the parent so we can query it, and removes one
record-cache-specific unit test that's not useful anymore.

Change-Id: Ifa0236a6f29f7633ceedce1300a7743c4f32105e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/914
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>