Rename demos to benchmarks, add README.md
This sets up a benchmarks directory which will include all code
used to benchmark ReadySet (and shared utilities). The benchmarks
are all refactored to fall under one crate allowing us to
build all the benchmarks with a single command:

cargo -Z unstable-options build -p benchmarks --out-dir \
        target/benchmarks/

This will pave the way for building an AMI with all the benchmark
code and data files.

Change-Id: I5a257764f05ec29851b9b6a7ca4669d8628e60d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/921
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>