Added replay path sorting to ensure we traverse them correctly.
We were handling replays incorrectly (in the wrong order), causing `Union`s to
return wrong results when more than one was present in the graph.
To understand what is going on, let me use an example.
Imagine you had two unions in a diamond shape:
a -> b -> u_1 -> d -> u_2 -> f
  -> c ->     -> e ->

There are going to be four different paths:
a -> b -> u_1 -> d -> u_2 -> f
a -> b -> u_1 -> e -> u_2 -> f
a -> c -> u_1 -> d -> u_2 -> f
a -> c -> u_1 -> e -> u_2 -> f

For the sake of making this easier to understand, let's represent those paths
based only on the node they choose at each bifurcation:
1. (b, d)
2. (b, e)
3. (c, d)
4. (c, e)

Now, before this `sort_by` was introduced, the paths were sorted like that.
The problem is that unions, when receiving replay messages, wait until they
receive all of them to release the full replay.
That waiting (buffering) mechanism relies solely on how many messages they get.
Once they get as many messages as ancestors they have, they release the full
replay.

Now, following the example, this is what happens:
1. u_1 receives message from b, and buffers it
2. u_1 receives message from b (which will be the same as before), now it has
two messages, so it releases it. WRONG.

So, the paths must be sorted based on the node indexes, but backwards
1. (b, d)
2. (c, d)
3. (b, e)
4. (c, e)

We now sort the paths to match that order.

As a final note, this is the _easiest_ fix I could think of, without having to
redesign how replay paths behave in the presence of `Union`s.
Imho, this solution is not ideal, since it adds another invariant to Noria which
might be hard to keep track of, and to maintain in the presence of changes.
It would be worth brainstorming a redesign for how replay paths work (or at
least when `Union`s are involved), because there's room for improvement as well.

Refs: ENG-653
Change-Id: Ic363eb19051026ace2af146bb21deba12835f39c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/918
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>